+++ BSUIR.CDOCM.ConstantMistake/Schema.cs
@@ -2,14 +2,14 @@
 {
     public abstract class Schema
     {
-        public int[] XVector { get; set; }
+        public byte[] XVector { get; set; }
         protected int[] Nodes { get; set; }
 
-        public abstract int F1(bool test);
-        public abstract int F2(bool test);
-        public abstract int F3(bool test);
-        public abstract int F4(bool test);
-        public abstract int F5(bool test);
-        public abstract int F6(bool test);
+        public abstract byte F1(bool test);
+        public abstract byte F2(bool test);
+        public abstract byte F3(bool test);
+        public abstract byte F4(bool test);
+        public abstract byte F5(bool test);
+        public abstract byte F6(bool test);
     }
 }
